# Precompilation for faster Pluto startup