@VarLad 2021-08-16
This is exciting!
So it'll only take 11secs to start Pluto, in exchange of larger precompile times?