@lungben 2021-08-17
It looks like (from the timings above) that about 20% of the startup time is due to compilations. Ideally we could be able to move this part mostly to precompilation. In the current state, this PR reduces precompilation times by about 2/3.
The main difficulty with the remaining 1/3 is that we cannot just call "standard" Pluto functions in the precompilation phase because they contain objects that cannot be serialized (especially Tasks).

The remaining 80% of the Pluto startup time seems not to be related to compilation times.