@fonsp 2021-08-22
Cool! I tried this but didn't manage, but it might still be useful to look at: https://github.com/fonsp/Pluto.jl/tree/precompile-during-precompile

I'll have more time to look at the PR after my break!