@lungben 2021-08-22
Maybe `workspace_use_distributed=true` prevented some precompilation?
On the flip side, `run_reactive!` crashes for me if called during precompilation, most likely due to task / async functionality (which is not allowed in precompilation) in WorkspaceManager.