@fonsp 2021-09-16
Can you time `using Pluto`?