@lungben 2021-09-16
**Current main:**

PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.773140 seconds (1.08 M allocations: 69.001 MiB, 3.10% gc time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.792773 seconds (1.08 M allocations: 69.002 MiB, 3.01% gc time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.830735 seconds (1.08 M allocations: 69.001 MiB, 3.61% gc time)

PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 34.839933 seconds (36.48 M allocations: 2.089 GiB, 2.14% gc time, 18.12% compilation time)
 10.727400 seconds (598.29 k allocations: 40.657 MiB, 0.11% gc time, 0.36% compilation time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 30.045399 seconds (36.48 M allocations: 2.090 GiB, 2.56% gc time, 21.94% compilation time)
 10.591936 seconds (594.17 k allocations: 40.401 MiB, 0.12% gc time, 0.38% compilation time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 31.160570 seconds (36.48 M allocations: 2.090 GiB, 2.53% gc time, 23.23% compilation time)
 10.726944 seconds (595.25 k allocations: 40.477 MiB, 0.12% gc time, 0.37% compilation time)

**With this PR:**

PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.965575 seconds (1.38 M allocations: 88.982 MiB, 2.54% gc time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.944487 seconds (1.38 M allocations: 88.982 MiB, 2.36% gc time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia -e '@time using Pluto'

  0.998248 seconds (1.38 M allocations: 88.980 MiB, 2.48% gc time)


PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 29.648578 seconds (40.87 M allocations: 2.339 GiB, 2.48% gc time, 7.97% compilation time)
 10.604080 seconds (699.69 k allocations: 43.801 MiB, 0.08% gc time, 0.39% compilation time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 33.155839 seconds (40.87 M allocations: 2.339 GiB, 2.42% gc time, 8.51% compilation time)
 11.402447 seconds (703.24 k allocations: 44.066 MiB, 0.08% gc time, 0.37% compilation time)
PS C:\Users\benlu\.julia\dev\Pluto\test> julia notebook_startup.jl

 31.700384 seconds (40.86 M allocations: 2.339 GiB, 2.34% gc time, 7.49% compilation time)
 10.511332 seconds (703.15 k allocations: 44.064 MiB, 0.09% gc time, 0.39% compilation time)


Looks like there are no consistent speed-ups with this PR (anymore?). Pure `using` timing seems even slightly worse. However, the compilation time fraction is still decreased by a factor of 3 (I wonder why this does not result in a faster startup anymore).